+++ commercetools-models/src/main/java/io/sphere/sdk/carts/TaxPortion.java
@@ -1,42 +1,25 @@
 package io.sphere.sdk.carts;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import javax.annotation.Nullable;
 import javax.money.MonetaryAmount;
 
-public final class TaxPortion extends Base {
-    private final Double rate;
-    private final MonetaryAmount amount;
-    @Nullable
-    private final String name;
+@JsonDeserialize(as = TaxPortionImpl.class)
+public interface TaxPortion {
 
-    @JsonCreator
-    private TaxPortion(final Double rate, final MonetaryAmount amount, final String name) {
-        this.rate = rate;
-        this.amount = amount;
-        this.name = name;
+    static TaxPortion of(final double rate, final MonetaryAmount amount) {
+        return new TaxPortionImpl(rate, amount, null);
     }
 
-    public static TaxPortion of(final double rate, final MonetaryAmount amount) {
-        return new TaxPortion(rate, amount, null);
+    static TaxPortion of(final double rate, final MonetaryAmount amount, final String name) {
+        return new TaxPortionImpl(rate, amount, name);
     }
 
-    public static TaxPortion of(final double rate, final MonetaryAmount amount, final String name) {
-        return new TaxPortion(rate, amount, name);
-    }
+    Double getRate();
 
-    public Double getRate() {
-        return rate;
-    }
-
-    public MonetaryAmount getAmount() {
-        return amount;
-    }
+    MonetaryAmount getAmount();
 
     @Nullable
-    public String getName() {
-        return name;
-    }
+    String getName();
 }
